initial impl of BusRideState

+++ Lazarus/src/com/android/lazarus/serviceadapter/BusReportingServiceAdapter.java
@@ -8,5 +8,5 @@ public interface BusReportingServiceAdapter {
 	 * 
 	 * @return bus with given id, null if none found
 	 */
-	public Bus getBus(String token, int busId);
+	public Bus getBus(String token, long busId);
 }
